Base: Set up module support, theme and gitignore
Set up the project to accept modules and pulled in the first module -
the nightfall theme by LordMathis. To get this theme to work, required
parameters were set in the `hugo.toml` file.

Also added `resources/_gen` and `.hugo_build.lock` to the gitignore
file. They are generated by Hugo and should not be commited to the
repository.