Index: Reformatted index page
Wrote an `index.html` file which overwrites the default one from the
Nightfall theme. It removes the vertical alignment from the `content`
section. A recent posts section was also added to the index page.

To support the recent post feature a new partial was created, which
lists the last three posts created for the blog.