Licensing: Added licenses
I've picked CC Attribution-ShareAlike 4.0 License for this blog.
Reference to this license has been added to the footer of the page. I've
also removed an unecessary section from the footer partial relating to
customized footers. This should have been a separate commit, but eh.

The full license text was also added to the root of this repository.

Creative Commons has been noted in the Contributors page for their
license and a tool to help me pick the license.