Actions: Created a build job
Created a job that runs in pull requests or on regular pushes. This job
builds the minified version if the site using Hugo.

After successfully creating the static website, a short lived artefact
(1 day) is uploaded. This is intended to be used to view the content
being built or to be used in subsequent jobs.